Move app to server package
