+++ server/app.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package temporal
+package server
 
 import (
 	"log"
@@ -38,7 +38,6 @@ import (
 	"go.temporal.io/server/common/log/loggerimpl"
 	"go.temporal.io/server/common/primitives"
 	"go.temporal.io/server/common/service/config"
-	"go.temporal.io/server/server"
 	"go.temporal.io/server/tools/cassandra"
 	"go.temporal.io/server/tools/sql"
 )
@@ -53,89 +52,8 @@ var (
 	}
 )
 
-func loadConfig(c *cli.Context) *config.Config {
-	env := strings.TrimSpace(c.GlobalString("env"))
-	zone := strings.TrimSpace(c.GlobalString("zone"))
-	configDir := getConfigDir(c)
-
-	log.Printf("Loading config; env=%v,zone=%v,configDir=%v\n", env, zone, configDir)
-
-	var cfg config.Config
-	err := config.Load(env, configDir, zone, &cfg)
-	if err != nil {
-		log.Fatal("Config file corrupted.", err)
-	}
-	if cfg.Log.Level == "debug" {
-		log.Printf("config=\n%v\n", cfg.String())
-	}
-
-	if err = cfg.Validate(); err != nil {
-		log.Fatalf("config validation failed: %v", err)
-	}
-	// cassandra schema version validation
-	if err = cassandra.VerifyCompatibleVersion(cfg.Persistence); err != nil {
-		log.Fatalf("cassandra schema version compatibility check failed: %v", err)
-	}
-	// sql schema version validation
-	if err = sql.VerifyCompatibleVersion(cfg.Persistence); err != nil {
-		log.Fatalf("sql schema version compatibility check failed: %v", err)
-	}
-
-	if err = cfg.Global.PProf.NewInitializer(loggerimpl.NewLogger(cfg.Log.NewZapLogger())).Start(); err != nil {
-		log.Fatalf("fail to start PProf: %v", err)
-	}
-
-	return &cfg
-}
-
-// getServices parses the services arg from cli
-// and returns a list of services to start
-func getServices(c *cli.Context) []string {
-
-	val := strings.TrimSpace(c.String("services"))
-	tokens := strings.Split(val, ",")
-
-	if len(tokens) == 0 {
-		log.Fatal("list of services is empty")
-	}
-
-	for _, t := range tokens {
-		if !isValidService(t) {
-			log.Fatalf("invalid service `%v` in service list [%v]", t, val)
-		}
-	}
-
-	return tokens
-}
-
-func isValidService(in string) bool {
-	for _, s := range validServices {
-		if s == in {
-			return true
-		}
-	}
-	return false
-}
-
-func getConfigDir(c *cli.Context) string {
-	return path.Join(getRootDir(c), c.GlobalString("config"))
-}
-
-func getRootDir(c *cli.Context) string {
-	dirpath := c.GlobalString("root")
-	if len(dirpath) == 0 {
-		cwd, err := os.Getwd()
-		if err != nil {
-			log.Fatalf("os.Getwd() failed, err=%v", err)
-		}
-		return cwd
-	}
-	return dirpath
-}
-
 // BuildCLI is the main entry point for the temporal server
 func BuildCLI() *cli.App {
-
 	app := cli.NewApp()
 	app.Name = "temporal"
 	app.Usage = "Temporal server"
@@ -184,13 +102,13 @@ func BuildCLI() *cli.App {
 				cfg := loadConfig(c)
 				services := getServices(c)
 
-				var servers []*server.Server
+				var servers []*Server
 				sigtermCh := make(chan os.Signal, 1)
 				signal.Notify(sigtermCh, os.Interrupt, syscall.SIGTERM)
 				for _, svc := range services {
-					s := server.New(
-						server.ForService(svc),
-						server.WithConfig(cfg),
+					s := New(
+						ForService(svc),
+						WithConfig(cfg),
 					)
 					servers = append(servers, s)
 					s.Start()
@@ -206,4 +124,83 @@ func BuildCLI() *cli.App {
 		},
 	}
 	return app
+}
+
+func loadConfig(c *cli.Context) *config.Config {
+	env := strings.TrimSpace(c.GlobalString("env"))
+	zone := strings.TrimSpace(c.GlobalString("zone"))
+	configDir := getConfigDir(c)
+
+	log.Printf("Loading config; env=%v,zone=%v,configDir=%v\n", env, zone, configDir)
+
+	var cfg config.Config
+	err := config.Load(env, configDir, zone, &cfg)
+	if err != nil {
+		log.Fatal("Config file corrupted.", err)
+	}
+	if cfg.Log.Level == "debug" {
+		log.Printf("config=\n%v\n", cfg.String())
+	}
+
+	if err = cfg.Validate(); err != nil {
+		log.Fatalf("config validation failed: %v", err)
+	}
+	// cassandra schema version validation
+	if err = cassandra.VerifyCompatibleVersion(cfg.Persistence); err != nil {
+		log.Fatalf("cassandra schema version compatibility check failed: %v", err)
+	}
+	// sql schema version validation
+	if err = sql.VerifyCompatibleVersion(cfg.Persistence); err != nil {
+		log.Fatalf("sql schema version compatibility check failed: %v", err)
+	}
+
+	if err = cfg.Global.PProf.NewInitializer(loggerimpl.NewLogger(cfg.Log.NewZapLogger())).Start(); err != nil {
+		log.Fatalf("fail to start PProf: %v", err)
+	}
+
+	return &cfg
+}
+
+// getServices parses the services arg from cli
+// and returns a list of services to start
+func getServices(c *cli.Context) []string {
+	val := strings.TrimSpace(c.String("services"))
+	tokens := strings.Split(val, ",")
+
+	if len(tokens) == 0 {
+		log.Fatal("list of services is empty")
+	}
+
+	for _, t := range tokens {
+		if !isValidService(t) {
+			log.Fatalf("invalid service `%v` in service list [%v]", t, val)
+		}
+	}
+
+	return tokens
+}
+
+func isValidService(in string) bool {
+	for _, s := range validServices {
+		if s == in {
+			return true
+		}
+	}
+	return false
+}
+
+func getConfigDir(c *cli.Context) string {
+	return path.Join(getRootDir(c), c.GlobalString("config"))
+}
+
+func getRootDir(c *cli.Context) string {
+	dirpath := c.GlobalString("root")
+	if len(dirpath) == 0 {
+		cwd, err := os.Getwd()
+		if err != nil {
+			log.Fatalf("os.Getwd() failed, err=%v", err)
+		}
+		return cwd
+	}
+	return dirpath
 }
